+++ oliver/lynx-tasm/ChangeLog.md
@@ -1,3 +1,6 @@
+# 0.0.6
+* support `decode_napi` to resolve a tasm file.
+
 # 0.0.5
 * Update supported lynx version to 3.2
 
@@ -5,4 +8,4 @@
 * Update supported lynx version to 3.0
 
 # 0.0.3
-* support compile `@lynx-dev/tasm` as wasm binary.
+* support compile `@lynx-js/tasm` as wasm binary.